+++ StockTaxSaver/StockTaxSaver/ViewModel/CalculatorElementViewModel.swift
@@ -16,12 +16,12 @@ class CalculatorElementViewModel {
     }
     
     func setPurchasePrice(price: String) {
-        let priceDouble = Double(price) ?? 0
+        let priceDouble = Int(price) ?? 0
         calculatorElement.purchasePrice = priceDouble
     }
     
     func setCurrentPrice(price: String) {
-        let priceDouble = Double(price) ?? 0
+        let priceDouble = Int(price) ?? 0
         calculatorElement.currentPrice = priceDouble
     }
 }
